# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"JsonObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "wish-list-backend",
		"version": "1.0.0",
		"description": "backend for wish-list",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "VH"
		}
	},
	"paths": {
		"/api/icecream/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/icecream/getAll": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/icecream/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"name",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/icecream/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/icecream/getToppings": {
			"get": {
				"operationId": "AllToppings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/icecream/fields": {
			"get": {
				"operationId": "Fields",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/icecream/query": {
			"get": {
				"operationId": "Query",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/drink/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"size": {
										"type": "number",
										"format": "double"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"size",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/task/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"description": {
										"type": "string"
									},
									"status": {
										"type": "string"
									}
								},
								"required": [
									"description",
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/task/getAll": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/task/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"description": {
										"type": "string"
									},
									"status": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"description",
									"status",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/task/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}